+++ tests/pod_hash_test.cpp
@@ -15,13 +15,9 @@ struct Data {
     uint64_t mp;
 };
 
-uint64_t Data_hash(const Data &d) {
-    return murmur_hash_64(&d, sizeof(Data), 0xDEADBEEF);
-}
+uint64_t Data_hash(const Data &d) { return murmur_hash_64(&d, sizeof(Data), 0xDEADBEEF); }
 
-bool Data_equal(const Data &d1, const Data &d2) {
-    return d1.id == d2.id && d1.hp == d2.hp && d1.mp == d2.mp;
-}
+bool Data_equal(const Data &d1, const Data &d2) { return d1.id == d2.id && d1.hp == d2.hp && d1.mp == d2.mp; }
 
 int main() {
     Data D1 = {100lu, 100lu, 100lu};
@@ -30,16 +26,15 @@ int main() {
     memory_globals::init();
 
     {
-        PodHash<Data, uint64_t> h(memory_globals::default_allocator(),
-                                  memory_globals::default_allocator(),
+        PodHash<Data, uint64_t> h(memory_globals::default_allocator(), memory_globals::default_allocator(),
                                   Data_hash, Data_equal);
 
         pod_hash::reserve(h, 512);
 
         assert(pod_hash::has(h, D1) == false);
-        pod_hash::set(h, D1, 0x10lu);
+        pod_hash::set(h, D1, uint64_t(0x10lu));
         assert(pod_hash::has(h, D1) == true);
-        assert(pod_hash::set_default(h, D2, 0x10lu) == 0x10lu);
+        assert(pod_hash::set_default(h, D2, uint64_t(0x10lu)) == 0x10lu);
 
         pod_hash::remove(h, D1);
         pod_hash::remove(h, D2);
@@ -51,7 +46,7 @@ int main() {
 
         for (uint64_t i = 0; i < 1000; ++i) {
             Data d = {i, i, i};
-            assert(pod_hash::set_default(h, d, 0lu) == i * i);
+            assert(pod_hash::set_default(h, d, uint64_t(0lu)) == i * i);
         }
 
         log_info("Max chain length: %i\n", pod_hash::max_chain_length(h));
@@ -83,8 +78,7 @@ int main() {
             assert(!pod_hash::has(h, d));
         }
 
-        PodHash<char, uint64_t> h1(memory_globals::default_allocator(),
-                                   memory_globals::default_allocator(),
+        PodHash<char, uint64_t> h1(memory_globals::default_allocator(), memory_globals::default_allocator(),
                                    usual_hash<char>, usual_equal<char>);
 
         for (char i = 'a'; i < 'z'; ++i) {
